requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
fastapi-mail==1.4.1
qrcode[pil]==7.4.2
python-multipart==0.0.6
slowapi==0.1.9